Add more test cases, some comments.
Move anonymous calls into azure-anonymous.go
